microbial-segmentation-app/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
torch
torchvision
opencv-python
pillow
numpy
scikit-image
pandas
aiofiles
